refactor: use usize for model layer linear layer definition

+++ dynamic_learned_index/benches/model_benchmark.rs
@@ -13,7 +13,7 @@ const VECTOR_DIM: usize = 768;
 const SAMPLE_SIZE: usize = 10000;
 const NUM_CLUSTERS: usize = 9;
 const BATCH_SIZE: i64 = 256;
-const HIDDEN_NEURONS: i64 = 256;
+const HIDDEN_NEURONS: usize = 256;
 const EPOCHS: usize = 1;
 
 fn generate_random_data(size: usize, dim: usize) -> Vec<f32> {
